fix: make hash bit size be configurable

+++ chord/finger.go
@@ -27,10 +27,10 @@ func NewFingerTable(id model.HashID) []*Finger {
 func NewFinger(id model.HashID, index int, successor RingNode) *Finger {
 	nodeID := big.NewInt(0).SetBytes(id)
 	base := big.NewInt(2)
-	offset := big.NewInt(0).Exp(base, big.NewInt(int64(index)), nil) // 2^i
-	sum := big.NewInt(0).Add(nodeID, offset)                         // n + 2^i
-	ring := big.NewInt(0).Exp(base, big.NewInt(model.BitSize), nil)  //2^m
-	fingerID := big.NewInt(0).Mod(sum, ring)                         // (n + 2^i) mod 2^m
+	offset := big.NewInt(0).Exp(base, big.NewInt(int64(index)), nil)       // 2^i
+	sum := big.NewInt(0).Add(nodeID, offset)                               // n + 2^i
+	ring := big.NewInt(0).Exp(base, big.NewInt(int64(model.BitSize)), nil) //2^m
+	fingerID := big.NewInt(0).Mod(sum, ring)                               // (n + 2^i) mod 2^m
 	return &Finger{
 		Index: index,
 		ID:    fingerID.Bytes(),
